+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 # Next Release
 
 * **Admin Fix**: Fix error in Importing under Postgres
+* **Admin Feature** Ability to add entity to LIS Added
+* Added code endpoint for LIS organisations for web app use
+* Schema updated to account for these changes
 
 # v0.9.6